Blue top spins in middle of image.
Here is a series of images (created every 1000 frames) from my
modification of the 5th sketch from week 3 of the FutureLearn /
Creative Coding course.

My rule modifications:

1. White spinning top is more opaque at edges of image, more
transparent in middle of image (I'm using the distance from the centre,
then mapping that to a percentage transparency between 0 and 20).

2. Blue spinning top has the opposite rules for transparency.  Also,
it's boundary/margin for &quot;turning back&quot; is 300 pixels,
instead of 100 pixels.